bot_app/app/utils/config.py
import os
from dotenv import load_dotenv

# Загрузка переменных окружения из файла .env
load_dotenv()

# Токен Telegram бота
BOT_TOKEN = os.getenv("BOT_TOKEN")

# Строка подключения к базе данных
DATABASE_URL = os.getenv("DATABASE_URL")